feat: BaseAuctioneer (#45)
* feat: auctioneer swapper
* fix: receiver check
* chore: remove receiver
* chore: constructor
* chore: rename
* chore: storage layout
* chore: interface
* chore: prettier
* chore: test BaseAuctioneer
* chore: oops
* chore: prettier
* fix: comment
* feat: prefix BaseAuctioneer members with auction
* feat: allow cooldown == auction length

+++ src/interfaces/Solidly/ISolidly.sol
@@ -9,15 +9,15 @@ interface ISolidly {
     }
 
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         route[] calldata routes,
         address to,
-        uint deadline
+        uint256 deadline
     ) external returns (uint256[] memory amounts);
 
     function getAmountsOut(
-        uint amountIn,
+        uint256 amountIn,
         route[] memory routes
     ) external view returns (uint256[] memory amounts);
 }
